+++ CHANGE_LOG.md
@@ -8,3 +8,6 @@
 - 【component-task】
 - 【component-config】
 - 【component-log】
+- 【erwin-component】JavaLib 3.0.1
+- 【erwin-component】api-result 3.0.0
+- 【erwin-component-common-mybatis-plus】PageUtils